move classes to a different package, add item support
