+++ src/main/java/com/redlimerl/detailab/api/DetailArmorBarAPI.java
@@ -1,6 +1,6 @@
 package com.redlimerl.detailab.api;
 
-import com.redlimerl.detailab.api.data.ArmorBarLoader;
+import com.redlimerl.detailab.data.ArmorBarLoader;
 import com.redlimerl.detailab.api.render.CustomArmorBar;
 import net.minecraft.item.ArmorItem;
 import net.minecraft.item.Item;
@@ -18,19 +18,19 @@ public class DetailArmorBarAPI {
 
     public static final ArmorBarLoader LOADER = new ArmorBarLoader();
     static HashMap<ArmorItem, CustomArmorBar> staticArmorList = new HashMap<>();
-    static HashMap<Item, CustomArmorBar> itemList = new HashMap<>();
+    static HashMap<Item, CustomArmorBar> staticItemList = new HashMap<>();
 
     public static Map<ArmorItem, CustomArmorBar> getArmorBarList() {
         return LOADER.getArmorList();
     }
     public static Map<Item, CustomArmorBar> getItemBarList() {
-        return itemList;
+        return LOADER.getItemList();
     }
 
     public static Map<ArmorItem, CustomArmorBar> getStaticArmorBarList() {
         return staticArmorList;
     }
     public static Map<Item, CustomArmorBar> getStaticItemBarList() {
-        return itemList;
+        return staticItemList;
     }
 }
