+++ src/main/resources/assets/detailab/armor_bar/elytra.json
@@ -1,0 +1,14 @@
+{
+  "type": "item",
+  "items": [ "minecraft:elytra" ],
+  "render": {
+    "texture": "detailab:textures/armor_bar.png",
+    "texture_width": 128,
+    "texture_height": 128,
+    "offsets": {
+      "full": [ 36, 0 ],
+      "outline": [ 54, 0 ]
+    },
+    "shown": true
+  }
+}
